Fix cloudscale bucket provisioning
After Cloudscale's Ceph upgrade on May 7th, the buckets started to
not provision properly.

After debugging with an engineer from Cloudscale we found a solution:
We simply don't explicitly set the region for the create bucket API call.

This solves the issue and will also still work if the underlying Ceph
issues got fixed. According to the Cloudscale engineer the region is
optional.